requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
google-genai==1.63.0
Pillow==11.1.0
langchain==0.3.25
langchain-google-genai==2.1.4
numpy==2.2.0
